Revert "Forms: move international phone input to production (#44922)"
This reverts commit f5ae5f9755d614598d780e7364eada36e47be4da.
