--- projects/packages/forms/changelog/move-international-phone-input-to-telephone
@@ -1,4 +1,0 @@
-Significance: minor
-Type: changed
-
-Forms: move all international phone code and UI back into legacy telephone field, keep backwards compat
