+++ projects/packages/forms/src/blocks/contact-form/child-blocks.js
@@ -14,6 +14,7 @@ import JetpackImageSelectField from '../field-image-select';
 import JetpackMultipleChoiceField from '../field-multiple-choice';
 import JetpackNameField from '../field-name';
 import JetpackNumberField from '../field-number';
+import JetpackPhoneField from '../field-phone';
 import JetpackRatingField from '../field-rating';
 import JetpackDropdownField from '../field-select';
 import JetpackSingleChoiceField from '../field-single-choice';
@@ -61,7 +62,9 @@ export const childBlocks = [
 	...( getJetpackBlocksVariation() === 'experimental'
 		? [ JetpackRatingField, JetpackRatingInput, JetpackFieldSlider, JetpackSliderInput ]
 		: [] ),
-	...( getJetpackBlocksVariation() === 'beta' ? [ JetpackTimeField, JetpackPhoneInput ] : [] ),
+	...( getJetpackBlocksVariation() === 'beta'
+		? [ JetpackTimeField, JetpackPhoneField, JetpackPhoneInput ]
+		: [] ),
 
 	// The following are required for these blocks to be parsed correctly in block
 	// deprecations. They have been flagged with `supports.inserter: false` to
