+++ projects/packages/forms/src/blocks/contact-form/class-contact-form-block.php
@@ -227,41 +227,6 @@ class Contact_Form_Block {
 			)
 		);
 
-		Blocks::jetpack_register_block(
-			'jetpack/phone-input',
-			array(
-				'supports'     => array(
-					'__experimentalBorder' => array(
-						'color'  => true,
-						'radius' => true,
-						'style'  => true,
-						'width'  => true,
-					),
-					'color'                => array(
-						'text'       => true,
-						'background' => true,
-						'gradients'  => false,
-					),
-					'typography'           => array(
-						'fontSize'                     => true,
-						'lineHeight'                   => true,
-						'__experimentalFontFamily'     => true,
-						'__experimentalFontWeight'     => true,
-						'__experimentalFontStyle'      => true,
-						'__experimentalTextTransform'  => true,
-						'__experimentalTextDecoration' => true,
-						'__experimentalLetterSpacing'  => true,
-					),
-				),
-				'uses_context' => array(
-					'jetpack/field-share-attributes',
-					'jetpack/field-prefix-options',
-					'jetpack/field-prefix-default',
-					'jetpack/field-prefix-onChange',
-				),
-			)
-		);
-
 		if ( Blocks::get_variation() === 'experimental' ) {
 			Blocks::jetpack_register_block(
 				'jetpack/input-rating',
@@ -295,6 +260,44 @@ class Contact_Form_Block {
 							'__experimentalTextDecoration' => true,
 							'__experimentalLetterSpacing'  => true,
 						),
+					),
+				)
+			);
+		}
+
+		if ( Blocks::get_variation() === 'beta' ) {
+
+			Blocks::jetpack_register_block(
+				'jetpack/phone-input',
+				array(
+					'supports'     => array(
+						'__experimentalBorder' => array(
+							'color'  => true,
+							'radius' => true,
+							'style'  => true,
+							'width'  => true,
+						),
+						'color'                => array(
+							'text'       => true,
+							'background' => true,
+							'gradients'  => false,
+						),
+						'typography'           => array(
+							'fontSize'                     => true,
+							'lineHeight'                   => true,
+							'__experimentalFontFamily'     => true,
+							'__experimentalFontWeight'     => true,
+							'__experimentalFontStyle'      => true,
+							'__experimentalTextTransform'  => true,
+							'__experimentalTextDecoration' => true,
+							'__experimentalLetterSpacing'  => true,
+						),
+					),
+					'uses_context' => array(
+						'jetpack/field-share-attributes',
+						'jetpack/field-prefix-options',
+						'jetpack/field-prefix-default',
+						'jetpack/field-prefix-onChange',
 					),
 				)
 			);
@@ -343,24 +346,7 @@ class Contact_Form_Block {
 			'jetpack/field-telephone',
 			array(
 				'render_callback'  => array( Contact_Form_Plugin::class, 'gutenblock_render_field_telephone' ),
-				'attributes'       => array(
-					'showCountrySelector' => array(
-						'type' => 'boolean',
-					),
-					'default'             => array(
-						'type' => 'string',
-						'role' => 'content',
-					),
-				),
-				'supports'         => array(
-					'interactivity' => true,
-				),
-				'provides_context' => array(
-					'jetpack/field-share-attributes'     => 'shareAttributes',
-					'jetpack/field-required'             => 'required',
-					'jetpack/field-prefix-default'       => 'default',
-					'jetpack/field-phone-country-toggle' => 'showCountrySelector',
-				),
+				'provides_context' => array( 'jetpack/field-required' => 'required' ),
 			)
 		);
 		Blocks::jetpack_register_block(
@@ -471,6 +457,31 @@ class Contact_Form_Block {
 				array(
 					'render_callback'  => array( Contact_Form_Plugin::class, 'gutenblock_render_field_time' ),
 					'provides_context' => array( 'jetpack/field-required' => 'required' ),
+				)
+			);
+
+			Blocks::jetpack_register_block(
+				'jetpack/field-phone',
+				array(
+					'render_callback'  => array( Contact_Form_Plugin::class, 'gutenblock_render_field_phone' ),
+					'attributes'       => array(
+						'showCountrySelector' => array(
+							'type'    => 'boolean',
+							'default' => true,
+						),
+						'default'             => array(
+							'type'    => 'string',
+							'default' => '',
+							'role'    => 'content',
+						),
+					),
+					'supports'         => array(
+						'interactivity' => true,
+					),
+					'provides_context' => array(
+						'jetpack/field-required'       => 'required',
+						'jetpack/field-prefix-default' => 'default',
+					),
 				)
 			);
 		}
