+++ projects/packages/forms/src/blocks/contact-form/editor.scss
@@ -314,7 +314,7 @@
 		word-break: normal;
 		color: unset;
 		opacity: 0.6;
-		font-size: 72%;
+		font-size: 85%;
 		margin-inline-start: 0.25em;
 		font-weight: 400;
 	}
