+++ projects/packages/forms/src/blocks/contact-form/util/form-styles.js
@@ -128,7 +128,6 @@ window.jetpackForms.generateStyleVariables = function ( formNode ) {
 		fontSize,
 		fontFamily,
 		lineHeight,
-		inputHeight,
 	} = window.getComputedStyle( inputNode );
 
 	styleProbe.remove();
@@ -149,7 +148,6 @@ window.jetpackForms.generateStyleVariables = function ( formNode ) {
 		'--jetpack--contact-form--input-padding': inputPadding,
 		'--jetpack--contact-form--input-padding-top': inputPaddingTop,
 		'--jetpack--contact-form--input-padding-left': inputPaddingLeft,
-		'--jetpack--contact-form--input-height': inputHeight,
 		'--jetpack--contact-form--font-size': fontSize,
 		'--jetpack--contact-form--font-family': fontFamily,
 		'--jetpack--contact-form--line-height': lineHeight,
