+++ projects/packages/forms/src/blocks/field-phone/index.js
@@ -3,18 +3,18 @@ import { __ } from '@wordpress/i18n';
 import { mobile } from '@wordpress/icons';
 import defaultSettings from '../shared/settings';
 import { getIconColor } from '../shared/util/block-icons';
-import deprecated from './deprecated';
 import edit from './edit';
 import save from './save';
 
-const name = 'field-telephone';
+const name = 'field-phone';
 const settings = {
 	...defaultSettings,
-	title: __( 'Phone number field', 'jetpack-forms' ),
+	title: __( 'International Phone Number', 'jetpack-forms' ),
 	keywords: [
 		__( 'Phone', 'jetpack-forms' ),
 		__( 'Cellular phone', 'jetpack-forms' ),
 		__( 'Mobile', 'jetpack-forms' ),
+		__( 'International', 'jetpack-forms' ),
 	],
 	description: __( 'Collect phone numbers from site visitors.', 'jetpack-forms' ),
 	icon: {
@@ -22,7 +22,31 @@ const settings = {
 		src: <Icon icon={ mobile } />,
 	},
 	edit,
-	deprecated,
+	supports: {
+		...defaultSettings.supports,
+		interactivity: true,
+	},
+	attributes: {
+		...defaultSettings.attributes,
+		showCountrySelector: {
+			type: 'boolean',
+			default: true,
+		},
+		default: {
+			type: 'string',
+			default: '',
+			role: 'content',
+		},
+		onChangeDefaultCountry: {
+			type: 'function',
+		},
+	},
+	providesContext: {
+		...defaultSettings.providesContext,
+		'jetpack/field-prefix-default': 'default',
+		'jetpack/field-phone-country-toggle': 'showCountrySelector',
+	},
+	allowedBlocks: [ 'jetpack/label', 'jetpack/phone-input' ],
 	save,
 	example: {
 		innerBlocks: [
@@ -33,10 +57,7 @@ const settings = {
 				},
 			},
 			{
-				name: 'jetpack/input',
-				attributes: {
-					type: 'tel',
-				},
+				name: 'jetpack/phone-input',
 			},
 		],
 	},
