+++ projects/packages/forms/src/blocks/field-phone/save.js
@@ -1,0 +1,6 @@
+import { useInnerBlocksProps } from '@wordpress/block-editor';
+
+export default () => {
+	const innerBlocksProps = useInnerBlocksProps.save();
+	return <div { ...innerBlocksProps } />;
+};
