+++ projects/packages/forms/src/blocks/input-phone/editor.scss
@@ -1,5 +1,4 @@
-.jetpack-contact-form .jetpack-field.jetpack-field-phone,
-.jetpack-contact-form .jetpack-field.jetpack-field-telephone {
+.jetpack-contact-form .jetpack-field.jetpack-field-phone {
 
 	.jetpack-field__input,
 	.jetpack-field__textarea {
@@ -8,9 +7,7 @@
 		border-radius: var(--jetpack--contact-form--border-radius);
 		border-width: var(--jetpack--contact-form--border-size);
 		font-size: var(--jetpack--contact-form--font-size);
-		line-height: var(--jetpack--contact-form--line-height);
-		// defaulting to white, it's gutenberg's default background for inputs
-		background-color: var(--jetpack--contact-form--input-background, #fff);
+		line-height: 2; // matches default input line-height
 
 		// TODO: Doesn't work well with next version of Gutenberg, should be removed
 		&.is-selected {
@@ -19,7 +16,7 @@
 		}
 
 		// these class is used to style input and textarea elements
-		.jetpack-field__input-element:not(.jetpack-field__input-prefix .jetpack-field__input-element) {
+		.jetpack-field__input-element {
 			font: inherit;
 			line-height: inherit;
 			border: 0;
@@ -27,12 +24,11 @@
 			outline: none;
 			background-color: transparent;
 			padding: 0;
-			letter-spacing: inherit;
 
 			flex: 1;
+			width: inherit;
 			box-sizing: border-box;
 			resize: none;
-			min-height: unset;
 
 			&::before,
 			&::after {
@@ -41,10 +37,6 @@
 
 			&:focus {
 				box-shadow: none;
-			}
-
-			&::placeholder {
-				opacity: 0.5;
 			}
 		}
 
@@ -81,7 +73,6 @@
 				max-width: 100%;
 				text-overflow: ellipsis;
 				white-space: nowrap;
-				min-height: unset;
 
 				&:where(select) {
 					// compensate spacing for downarrow
@@ -97,8 +88,7 @@
 	}
 }
 
-.is-style-animated .jetpack-contact-form .jetpack-field.jetpack-field-phone,
-.is-style-animated .jetpack-contact-form .jetpack-field.jetpack-field-telephone {
+.is-style-animated .jetpack-contact-form .jetpack-field.jetpack-field-phone {
 
 	&:not(.has-placeholder):not(.is-selected),
 	&:not(.has-placeholder):not(.has-child-selected) {
@@ -117,8 +107,7 @@
 	}
 }
 
-.is-style-outlined .jetpack-contact-form .jetpack-field.jetpack-field-phone,
-.is-style-outlined .jetpack-contact-form .jetpack-field.jetpack-field-telephone {
+.is-style-outlined .jetpack-contact-form .jetpack-field.jetpack-field-phone {
 
 	&:not(.has-placeholder):not(.is-selected),
 	&:not(.has-placeholder):not(.has-child-selected) {
