+++ projects/packages/forms/src/contact-form/class-contact-form-plugin.php
@@ -507,13 +507,9 @@ class Contact_Form_Plugin {
 					continue;
 				}
 
-				// This input is exclusively used by the new telephone field.
+				// This input is exclusively used by the new phone field (not telephone).
 				if ( 'jetpack/phone-input' === $block_name ) {
 					$atts['placeholder'] = $inner_block['attrs']['placeholder'] ?? '';
-
-					if ( ! isset( $atts['showCountrySelector'] ) || ! $atts['showCountrySelector'] ) {
-						unset( $atts['default'] );
-					}
 
 					$input_attrs           = self::get_block_support_classes_and_styles( $block_name, $inner_block['attrs'] );
 					$atts['inputclasses']  = 'wp-block-jetpack-input jetpack-field__input-element';
@@ -1111,9 +1107,21 @@ class Contact_Form_Plugin {
 	 * @return string HTML for the contact form field.
 	 */
 	public static function gutenblock_render_field_telephone( $atts, $content, $block ) {
-		// conversion telephone to phone
-		$type = empty( $atts['showCountrySelector'] ) ? 'telephone' : 'phone';
-		$atts = self::block_attributes_to_shortcode_attributes( $atts, $type, $block );
+		$atts = self::block_attributes_to_shortcode_attributes( $atts, 'telephone', $block );
+		return Contact_Form::parse_contact_field( $atts, $content, $block );
+	}
+
+	/**
+	 * Render the phone field.
+	 *
+	 * @param array    $atts - the block attributes.
+	 * @param string   $content - html content.
+	 * @param WP_Block $block - the block instance object.
+	 *
+	 * @return string HTML for the contact form field.
+	 */
+	public static function gutenblock_render_field_phone( $atts, $content, $block ) {
+		$atts = self::block_attributes_to_shortcode_attributes( $atts, 'phone', $block );
 		return Contact_Form::parse_contact_field( $atts, $content, $block );
 	}
